Develop

+++ src/main/java/com/devee/devhive/domain/project/controller/ProjectController.java
@@ -1,13 +1,17 @@
 package com.devee.devhive.domain.project.controller;
 
+import static com.devee.devhive.global.exception.ErrorCode.PLEASE_CHANGE_NICKNAME;
 import static com.devee.devhive.global.exception.ErrorCode.PROJECT_CANNOT_DELETED;
+import static com.devee.devhive.global.exception.ErrorCode.UNAUTHORIZED;
 
 import com.devee.devhive.domain.project.apply.service.ProjectApplyService;
-import com.devee.devhive.domain.project.comment.reply.service.ReplyService;
+import com.devee.devhive.domain.project.chat.entity.ProjectChatRoom;
+import com.devee.devhive.domain.project.chat.service.ChatMemberService;
+import com.devee.devhive.domain.project.chat.service.ChatMessageService;
+import com.devee.devhive.domain.project.chat.service.ChatRoomService;
 import com.devee.devhive.domain.project.comment.service.CommentService;
 import com.devee.devhive.domain.project.entity.Project;
 import com.devee.devhive.domain.project.entity.dto.CreateProjectDto;
-import com.devee.devhive.domain.project.entity.dto.ProjectImageDto;
 import com.devee.devhive.domain.project.entity.dto.ProjectInfoDto;
 import com.devee.devhive.domain.project.entity.dto.ProjectListDto;
 import com.devee.devhive.domain.project.entity.dto.SearchProjectDto;
@@ -19,21 +23,28 @@ import com.devee.devhive.domain.project.member.service.ProjectMemberService;
 import com.devee.devhive.domain.project.service.ProjectService;
 import com.devee.devhive.domain.project.techstack.service.ProjectTechStackService;
 import com.devee.devhive.domain.project.type.ApplyStatus;
+import com.devee.devhive.domain.project.type.ProjectStatus;
+import com.devee.devhive.domain.project.views.service.ViewCountService;
 import com.devee.devhive.domain.techstack.entity.dto.TechStackDto;
+import com.devee.devhive.domain.user.bookmark.entity.Bookmark;
 import com.devee.devhive.domain.user.bookmark.service.BookmarkService;
 import com.devee.devhive.domain.user.entity.User;
 import com.devee.devhive.domain.user.entity.dto.SimpleUserDto;
 import com.devee.devhive.domain.user.favorite.service.FavoriteService;
 import com.devee.devhive.domain.user.service.UserService;
 import com.devee.devhive.domain.user.techstack.service.UserTechStackService;
+import com.devee.devhive.domain.user.type.Role;
 import com.devee.devhive.global.entity.PrincipalDetails;
 import com.devee.devhive.global.exception.CustomException;
 import com.devee.devhive.global.s3.S3Service;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
@@ -41,6 +52,7 @@ import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -53,15 +65,16 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/projects")
+@Tag(name = "PROJECT API", description = "프로젝트 API")
 public class ProjectController {
 
   private final UserService userService;
   private final ProjectService projectService;
   private final CommentService commentService;
-  private final ReplyService replyService;
   private final ProjectTechStackService projectTechStackService;
   private final ProjectMemberService projectMemberService;
   private final FavoriteService favoriteService;
@@ -69,16 +82,25 @@ public class ProjectController {
   private final BookmarkService bookmarkService;
   private final ProjectApplyService projectApplyService;
   private final S3Service s3Service;
+  private final ViewCountService viewCountService;
+  private final ChatRoomService chatRoomService;
+  private final ChatMemberService chatMemberService;
+  private final ChatMessageService chatMessageService;
 
   // 프로젝트 작성
   @PostMapping
+  @Operation(summary = "프로젝트 생성")
   public ResponseEntity<SimpleProjectDto> createProject(
       @AuthenticationPrincipal PrincipalDetails principal,
       @RequestBody @Valid CreateProjectDto createProjectDto
   ) {
     User user = userService.getUserByEmail(principal.getEmail());
+    if (user.getNickName().startsWith("닉네임변경해주세요")) {
+      throw new CustomException(PLEASE_CHANGE_NICKNAME);
+    }
 
     Project project = projectService.createProject(createProjectDto, user);
+    viewCountService.create(project);
     List<TechStackDto> techStacks = createProjectDto.getTechStacks();
     projectTechStackService.createProjectTechStacks(project, techStacks);
     projectMemberService.saveProjectLeader(user, project);
@@ -91,9 +113,10 @@ public class ProjectController {
 
   // 상태변경
   @PutMapping("/{projectId}/status")
+  @Operation(summary = "프로젝트 상태 변경", description = "프로젝트 고유 ID로 프로젝트 상태 변경 - 글 작성자만 변경 가능(모집중, 모집 완료, 재모집, 완료)")
   public void updateProjectStatus(
       @AuthenticationPrincipal PrincipalDetails principal,
-      @PathVariable Long projectId, @RequestBody UpdateProjectStatusDto statusDto
+      @PathVariable(name = "projectId") Long projectId, @RequestBody UpdateProjectStatusDto statusDto
   ) {
     User user = userService.getUserByEmail(principal.getEmail());
     List<ProjectMember> members = projectMemberService.getProjectMemberByProjectId(projectId);
@@ -102,9 +125,10 @@ public class ProjectController {
 
   // 프로젝트 수정
   @PutMapping("/{projectId}")
+  @Operation(summary = "프로젝트 수정", description = "프로젝트 고유 ID로 프로젝트 수정 - 글 작성자만 수정 가능")
   public void updateProject(
       @AuthenticationPrincipal PrincipalDetails principal,
-      @PathVariable Long projectId, @RequestBody @Valid UpdateProjectDto updateProjectDto
+      @PathVariable(name = "projectId") Long projectId, @RequestBody @Valid UpdateProjectDto updateProjectDto
   ) {
     User user = userService.getUserByEmail(principal.getEmail());
     Project project = projectService.updateProject(user, projectId, updateProjectDto);
@@ -113,23 +137,35 @@ public class ProjectController {
 
   // 프로젝트 삭제
   @DeleteMapping("/{projectId}")
+  @Operation(summary = "작성자가 프로젝트 삭제", description = "프로젝트 고유 ID로 프로젝트 삭제 - 글 작성자만 삭제 가능")
   public void deleteProject(
-      @AuthenticationPrincipal PrincipalDetails principal, @PathVariable Long projectId
+      @AuthenticationPrincipal PrincipalDetails principal, @PathVariable(name = "projectId") Long projectId
   ) {
     User user = userService.getUserByEmail(principal.getEmail());
     Project project = projectService.findById(projectId);
     if (!Objects.equals(project.getUser().getId(), user.getId())) {
       throw new CustomException(PROJECT_CANNOT_DELETED);
     }
+    if (project.getStatus() != ProjectStatus.RECRUITING) {
+      throw new CustomException(PROJECT_CANNOT_DELETED);
+    }
 
-    List<Long> commentIdList = commentService.deleteCommentsByProjectId(projectId);
-    replyService.deleteRepliesByCommentList(commentIdList);
-    projectTechStackService.deleteProjectTechStacksByProjectId(projectId);
-    projectMemberService.deleteProjectMembers(projectId);
+    deleteOfProject(projectId);
     projectService.deleteProject(project);
   }
 
+  // 리더가 퇴출되어 예외없이 프로젝트 삭제
+  @Transactional
+  @DeleteMapping("/{projectId}/leader-exit")
+  @Operation(summary = "리더가 퇴출되어 프로젝트 삭제", description = "프로젝트 고유 ID로 프로젝트 삭제")
+  public void deleteProject(@PathVariable(name = "projectId") Long projectId) {
+    Project project = projectService.findById(projectId);
+    deleteOfProject(projectId);
+    projectService.deleteLeadersProject(project);
+  }
+
   @PostMapping("/list")
+  @Operation(summary = "프로젝트 목록 조회", description = "키워드(제목,내용), 백/프론트, 온/오프라인, 테크스택 고유 ID 리스트")
   public ResponseEntity<Page<ProjectListDto>> getProjects(
       @RequestBody(required = false) SearchProjectDto searchRequest,
       @RequestParam(defaultValue = "desc") String sort, Pageable pageable
@@ -150,39 +186,62 @@ public class ProjectController {
           .map(projectMember -> SimpleUserDto.from(projectMember.getUser()))
           .collect(Collectors.toList());
 
-      boolean bookmarked = isLoggedInUserBookmark(user, project.getId());
+      Long bookmarkId = isLoggedInUserBookmark(user, project.getId());
 
-      return ProjectListDto.of(project, techStackDtoList, projectMemberDtoList, bookmarked);
+      return ProjectListDto.of(project, techStackDtoList, projectMemberDtoList, bookmarkId);
     });
 
     return ResponseEntity.ok(projectListDtoPage);
   }
 
   // 이미지 업로드 후 url 얻는 api
-  @GetMapping("/image")
-  public ResponseEntity<ProjectImageDto> getImageUrl(
+  @PostMapping("/image")
+  public String getImageUrl(
+      @AuthenticationPrincipal PrincipalDetails principal,
       @RequestPart(value = "image", required = false) MultipartFile multipartFile
   ) {
-    return ResponseEntity.ok(ProjectImageDto.from(s3Service.upload(multipartFile)));
+    User user = userService.getUserByEmail(principal.getEmail());
+    if (!(user.getRole() == Role.ADMIN || user.getRole() == Role.USER)) {
+      throw new CustomException(UNAUTHORIZED);
+    }
+    return s3Service.upload(multipartFile);
   }
 
   // 프로젝트 상세 조회
   @GetMapping("/{projectId}")
-  public ResponseEntity<ProjectInfoDto> getProjectInfo(@PathVariable("projectId") Long projectId) {
+  @Operation(summary = "프로젝트 상세 조회", description = "프로젝트 고유 ID로 프로젝트 조회")
+  public ResponseEntity<ProjectInfoDto> getProjectInfo(@PathVariable(name = "projectId") Long projectId) {
     Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
-    Project project = projectService.updateViewPoint(projectId);
+    Project project = projectService.findById(projectId);
+    viewCountService.incrementViewCount(project);
     List<TechStackDto> techStacks = getTechStacks(projectId);
 
     List<SimpleUserDto> projectMembers = getProjectMembers(projectId);
 
     User loggedInUser = getLoggedInUser(authentication);
-    boolean isBookmark = isLoggedInUserBookmark(loggedInUser, projectId);
+    Long bookmarkId = isLoggedInUserBookmark(loggedInUser, projectId);
     ApplyStatus applyStatus = getApplyStatus(loggedInUser, project);
 
     return ResponseEntity.ok(ProjectInfoDto.of(
-        project, techStacks, projectMembers, loggedInUser, isBookmark, applyStatus)
+        project, techStacks, projectMembers, loggedInUser, bookmarkId, applyStatus)
     );
+  }
+
+  private void deleteOfProject(Long projectId) {
+    ProjectChatRoom chatRoom = chatRoomService.findByProjectId(projectId);
+    if (chatRoom != null) {
+      Long chatRoomId = chatRoom.getId();
+      chatMessageService.deleteOfChatRoom(chatRoomId);
+      chatMemberService.deleteOfChatRoom(chatRoomId);
+      chatRoomService.deleteChatRoom(chatRoom);
+    }
+    bookmarkService.deleteByProject(projectId);
+    commentService.deleteCommentsByProjectId(projectId);
+    projectTechStackService.deleteProjectTechStacksByProjectId(projectId);
+    projectMemberService.deleteProjectMembers(projectId);
+    projectApplyService.deleteAll(projectId);
+    viewCountService.delete(projectId);
   }
 
   private List<TechStackDto> getTechStacks(Long projectId) {
@@ -205,8 +264,15 @@ public class ProjectController {
     return null;
   }
 
-  private boolean isLoggedInUserBookmark(User loggedInUser, Long projectId) {
-    return loggedInUser != null && bookmarkService.isBookmark(loggedInUser.getId(), projectId);
+  private Long isLoggedInUserBookmark(User loggedInUser, Long projectId) {
+    Long bookmarkId = null;
+    if (loggedInUser != null) {
+      Bookmark bookmark = bookmarkService.findByUserIdAndProjectId(loggedInUser.getId(), projectId);
+      if (bookmark != null) {
+        bookmarkId = bookmark.getId();
+      }
+    }
+    return bookmarkId;
   }
 
   private ApplyStatus getApplyStatus(User loggedInUser, Project project) {
